index에 로직 구현 제거
